Add 'es' locale
- Basic support for Spanish language has been added (es_MX). Even though
the translation was done under a certain regionalism of the language, I
tried to make the translation as general as possible. However, other
contributors can use it as a basis for adapting it to other regionalism
of the Spanish (e.g. 'es_ES', 'es_AR'), if they deem it appropriate.